Allow to-do item to be deleted
This adds a button to delete to-do items, and also adds an ID to each
item.

Previously, index was being used for the item key. This can cause
rendering issues when an item is deleted, as an existing element could
have the same index as the deleted index had.
This ensures that the correct item is rendered after a delete.