Fix "Add List" option ignoring clicks (#72)
Option was ignoring clicks between the icon and the text. This was
caused by an onClick function being defined on both the text and the
icon for adding a list. This changes that to instead use an update
listener on the menu to check when the value is changed to the "Add
List" option.

Fixes #61