squash! Use layout instead of undocumented scrollbar
Set height on list in homepage when it's too long

It seems like Naive UI has a problem with `max-height` when using
`n-scrollbar`. By setting the `height` when the list is too long, this
is mitigated. Key includes whether or not the list is long in order to
force re-render (otherwise height was only added on refresh). Null check
included in layout ref because this seems to make it a null value one
time during the "short" to "long" transition.